@@ -1,27 +1,49 @@
 {
   "name": "@sone-dao/tone-react-user-controller",
   "version": "0.6.3",
-  "repository": "https://github.com/sone-dao/tone-react-user-controller.git",
+  "main": "dist/cjs/src/index.js",
+  "module": "dist/esm/src/index.js",
+  "types": "dist/index.d.ts",
+  "repository": "https://github.com/sone-dao/tone-template-react.git",
+  "author": "Sone DAO <dev@sone.works>",
   "license": "MIT",
   "scripts": {
+    "prebuild": "rm -rf dist",
     "build": "rollup -c",
-    "dev": "node ./dev.js"
+    "lint": "yarn run eslint --ignore-path .gitignore --fix src/**/*",
+    "format": "prettier --write src"
   },
+  "files": [
+    "dist"
+  ],
   "devDependencies": {
+    "@commitlint/cli": "^17.6.3",
+    "@commitlint/config-conventional": "^17.6.3",
     "@rollup/plugin-commonjs": "^24.0.1",
     "@rollup/plugin-node-resolve": "^15.0.1",
     "@rollup/plugin-terser": "^0.4.0",
     "@rollup/plugin-typescript": "^11.0.0",
+    "@sone-dao/sone-react-utils": "^1.0.1",
+    "@sone-dao/tone-react-api": "^0.3.0",
     "@types/node": "^18.11.18",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
-    "chokidar": "^3.5.3",
+    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "@typescript-eslint/parser": "^5.59.7",
+    "eslint": "^8.41.0",
+    "eslint-plugin-react": "^7.32.2",
+    "husky": "^8.0.3",
+    "next": "^13.4.3",
     "postcss": "^8.4.21",
     "prettier": "^2.8.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rollup": "^3.12.1",
+    "rollup-plugin-analyze": "^0.3.0",
     "rollup-plugin-dts": "^5.1.1",
     "rollup-plugin-peer-deps-external": "^2.2.4",
     "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-preserve-directives": "^0.1.1",
     "sass": "^1.58.0",
     "shelljs": "^0.8.5",
     "tslib": "^2.5.0",
@@ -30,14 +52,8 @@
   "peerDependencies": {
     "@sone-dao/sone-react-utils": "^1.0.1",
     "@sone-dao/tone-react-api": "^0.3.0",
-    "next": "^13.1.6",
+    "next": "^13.4.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
-  },
-  "files": [
-    "dist"
-  ],
-  "main": "dist/cjs/index.js",
-  "module": "dist/esm/index.js",
-  "types": "dist/index.d.ts"
+  }
 }
