@@ -5,35 +5,63 @@ import typescript from '@rollup/plugin-typescript'
 import dts from 'rollup-plugin-dts'
 import PeerDepsExternalPlugin from 'rollup-plugin-peer-deps-external'
 import postcss from 'rollup-plugin-postcss'
-import packageJson from './package.json' assert { type: 'json' }
+import preserveDirectives from 'rollup-plugin-preserve-directives'
+
+const plugins = [
+  postcss({
+    extract: false,
+    modules: true,
+    use: ['sass'],
+  }),
+  commonjs(),
+  PeerDepsExternalPlugin(),
+  resolve(),
+  preserveDirectives.default(),
+  terser(),
+]
 
 export default [
   {
     input: 'src/index.ts',
     output: [
       {
-        file: packageJson.main,
+        dir: 'dist/cjs',
         format: 'cjs',
-        sourcemap: true,
-      },
-      {
-        file: packageJson.module,
-        format: 'esm',
+        interop: 'auto',
+        preserveModules: true,
         sourcemap: true,
       },
     ],
     plugins: [
-      postcss({
-        extract: false,
-        modules: true,
-        use: ['sass'],
-      }),
-      commonjs(),
-      PeerDepsExternalPlugin(),
-      resolve(),
-      terser(),
-      typescript({ tsconfig: './rollup.tsconfig.json' }),
+      ...plugins,
+      typescript({ tsconfig: './rollup.tsconfig.cjs.json' }),
     ],
+    onwarn(warning, warn) {
+      if (warning.code !== 'MODULE_LEVEL_DIRECTIVE') {
+        warn(warning)
+      }
+    },
+  },
+  {
+    input: 'src/index.ts',
+    output: [
+      {
+        dir: 'dist/esm',
+        format: 'esm',
+        interop: 'auto',
+        preserveModules: true,
+        sourcemap: true,
+      },
+    ],
+    plugins: [
+      ...plugins,
+      typescript({ tsconfig: './rollup.tsconfig.esm.json' }),
+    ],
+    onwarn(warning, warn) {
+      if (warning.code !== 'MODULE_LEVEL_DIRECTIVE') {
+        warn(warning)
+      }
+    },
   },
   {
     input: 'dist/esm/types/index.d.ts',
